==> notebooks/02_sentencias_procesamiento.ipynb <==
import re
from striprtf.striprtf import rtf_to_text
import csv

## Load files

def leer_archivo_rtf(ruta_archivo):
    try:
        with open(ruta_archivo, 'r', encoding='utf-8') as archivo:
            contenido_rtf = archivo.read()
            contenido_texto = rtf_to_text(contenido_rtf)
            return contenido_texto
    except Exception as e:
        print(f"Error al leer el archivo .rtf: {e}")
        return None

### Parce sentences

#### Sentence C-318

path = r'D:\0_Respaldo\000_KaggleX\Projects\Rag-Ley675\675_LAW\data\raw\Sentencias\C-318-02.rtf'
c_318 = leer_archivo_rtf(path)

#Creo la lista de nodos
lista_sentencias = []

patron_sentencia_titulo = r"(Sentencia\s\w-\d{3}/\d{2})"
patron_articulos_demandados = re.compile(r'II\.\s*NORMAS\s*DEMANDADAS\.\s*(.*?)(?=\nIII\. LA DEMANDA)', re.DOTALL)
patron_demanda = re.compile(r'(II\.\s*LA DEMANDA)\.\s*(.*?)(?=\nIV\. INTERVENCIÓN)', re.DOTALL)
patron_decision = re.compile(r'(VII\.\s*DECISION)\.\s*(.*?)(?=\nNotifíquese, comuníquese, publíquese,)', re.DOTALL)
patron_articulos = re.compile(r'(Artículo\s)(\d{1,2})', re.DOTALL)


patron_articulos_demandados

sentencia_n = re.search(patron_sentencia_titulo, c_318)

sentencia_n.group(1)

s_demanda = re.search(patron_demanda, c_318)
s_decision = re.search(patron_decision, c_318)
a_demandados = re.search(patron_articulos_demandados, c_318)
articulos_318 = re.findall(patron_articulos, a_demandados.group(1))

articulos_318

# Nodo
nodo_dict = {}
nodo_dict = {
    "sentencia_id":318,
    "sentencia_name":sentencia_n.group(1),
    "demanda": s_demanda.group(2),
    "decision":s_decision.group(2)
    }


lista_sentencias.append(nodo_dict)

lista_sentencias

#### Read C-738-02 file

path = r'D:\0_Respaldo\000_KaggleX\Projects\Rag-Ley675\675_LAW\data\raw\Sentencias\C-738-02.rtf'

c_738 = leer_archivo_rtf(path)

print(c_738[:50])

#Patron
patron_sentencia_titulo = r"(Sentencia\s\w-\d{3}/\d{2})"
patron_articulos_demandados = re.compile(r'EL\s*TEXTO\s*DE\s*LAS\s*NORMAS\s*ACUSADAS\s*([\s\S]*?)(?=\nLA\s*DEMANDA)', re.DOTALL)
patron_demanda = re.compile(r'(\nLA\s*DEMANDA\n)\s*(.*?)(?=INTERVENCIONES)', re.DOTALL)
patron_decision = re.compile(r'DECISIÓN\n\s*(.*?)(?=\nCópiese, notifíquese, publíquese)', re.DOTALL)
patron_articulos = re.compile(r'(Artículo\s)(\d{1,2})', re.DOTALL)

s_demanda = re.search(patron_demanda, c_738)
s_decision = re.search(patron_decision, c_738)
a_demandados = re.search(patron_articulos_demandados, c_738)
articulos_738 = re.findall(patron_articulos, a_demandados.group(1))

articulos_738

nodo_dict = {}
nodo_dict = {
    "sentencia_id":738,
    "sentencia_name":sentencia_n.group(1),
    "demanda": s_demanda.group(2),
    "decision":s_decision.group(1)
    }


lista_sentencias.append(nodo_dict)
lista_sentencias

#### Sentences 782

path = r'D:\0_Respaldo\000_KaggleX\Projects\Rag-Ley675\675_LAW\data\raw\Sentencias\C-782-04.rtf'

c_782 = leer_archivo_rtf(path)

#Patron
patron_sentencia_titulo = r"(Sentencia\s\w-\d{3}/\d{2})"
patron_demanda = re.compile(r'(LA\s*DEMANDA\.\n)\s*(.*?)(?=INTERVENCIONES)', re.DOTALL)
patron_decision = re.compile(r'DECISIÓN\n\s*(.*?)(?=\nNotifíquese, comuníquese, cúmplase, publíquese,)', re.DOTALL)


sentencia_n = re.search(patron_sentencia_titulo, c_782)
s_demanda = re.search(patron_demanda, c_782)
s_decision = re.search(patron_decision, c_782)
articulos_782 = [('Artículo ', '3'), ('Artículo ', '25')]

nodo_dict = {}
nodo_dict = {
    "sentencia_id":782,
    "sentencia_name":sentencia_n.group(1),
    "demanda": s_demanda.group(2),
    "decision":s_decision.group(1)
    }

lista_sentencias.append(nodo_dict)
lista_sentencias

#### Sentence SU509

path = r'D:\0_Respaldo\000_KaggleX\Projects\Rag-Ley675\675_LAW\data\raw\Sentencias\SU509-01.rtf'
su509 = leer_archivo_rtf(path)

print(su509[:25])

#Patron
patron_sentencia_titulo = r"(Sentencia\s[A-Z]{2}\.\d{3}/\d{2})"
patron_decision = re.compile(r'DECISION\n\s*(.*?)(?=\nCópiese, notifíquese, comuníquese)', re.DOTALL)
patron_antecedentes = re.compile(r'ANTECEDENTES\n\s*(.*?)(?=\nSENTENCIAS)', re.DOTALL)

s_ant = re.search(patron_antecedentes, su509)

sentencia_n = re.search(patron_sentencia_titulo, su509)
s_decision = re.search(patron_decision, su509)


nodo_dict = {}
nodo_dict = {
    "sentencia_id":509,
    "sentencia_name":sentencia_n.group(1),
    "demanda": "Reglas jurisprudenciales cuando hay suspensión de algunos servicios de administración en conjuntos de vivienda.",
    "decision":s_decision.group(1)
    }

lista_sentencias.append(nodo_dict)
lista_sentencias

### Tutelas    -03

#### t_315866

#Patrones tutelas
patron_t_315866 = re.compile(r'T-315866\n\s*(.*?)(?=-Resolución )', re.DOTALL)
ts_1 = re.compile(r'(\nT-315866:\s*el\s*fallo.*?)(?=\nT-315884:)', re.DOTALL)

t_1_hechos = re.search(patron_t_315866, s_ant.group(1))
t_1_resolucion = re.search(ts_1, su509)

#### t_315884

patron_t_315884 = re.compile(r'T-315884\n\s*(.*?)(?=\n2.3. Pruebas)', re.DOTALL)
ts_2 = re.compile(r'(\nT-315884:\s*el\s*fallo.*?)(?=\nT-\s*362549)', re.DOTALL)

t_2_hechos = re.search(patron_t_315884, s_ant.group(1))
t_2_resolucion = re.search(ts_2, su509)

#### t_362549

patron_t_362549 = re.compile(r'T-362549\n\s*(.*?)(?=\n3.3. Pruebas)', re.DOTALL)
ts_3 = re.compile(r'(\nT-\s*362549:\s*el\s*fallo.*?)(?=\nT-\s*363900)', re.DOTALL)

t_3_hechos = re.search(patron_t_362549, s_ant.group(1))
t_3_resolucion = re.search(ts_3, su509)

#### T_363900


patron_t_363900= re.compile(r'T-363900\n\s*(.*?)(?=\n4.3. Pruebas)', re.DOTALL)
ts_4 = re.compile(r'(\nT-\s*363900:\s*decidió\s*en.*?)(?=\nT-372830)', re.DOTALL)

t_4_hechos = re.search(patron_t_363900, s_ant.group(1))
t_4_resolucion = re.search(ts_4, su509)

#### t_372830

patron_t_372830= re.compile(r'T-372830\n\s*(.*?)(?=\n5.3. Pruebas)', re.DOTALL)
ts_5 = re.compile(r'(\nT-372830:\s*el\s*fallo.*?)(?=\nCONSIDERACIONES)', re.DOTALL)

t_5_hechos = re.search(patron_t_372830, s_ant.group(1))
t_5_resolucion = re.search(ts_5, su509)

tutelas_list = []
tutelas_list = [{
    "tutela_id":315866,
    "tutela_name":'T-315866',
    "demanda":t_1_hechos.group(1),
    "decision":t_1_resolucion.group(1)
    },
{
    "tutela_id":315884,
    "tutela_name":'T-315884',
    "demanda":t_2_hechos.group(1),
    "decision":t_2_resolucion.group(1)
    },
{
    "tutela_id":362549,
    "tutela_name":'T-362549',
    "demanda":t_3_hechos.group(1),
    "decision":t_3_resolucion.group(1)
    },
{
    "tutela_id":363900,
    "tutela_name":'T-363900',
    "demanda":t_4_hechos.group(1),
    "decision":t_4_resolucion.group(1)
    },
{
    "tutela_id":372830,
    "tutela_name":'T-372830',
    "demanda":t_5_hechos.group(1),
    "decision":t_5_resolucion.group(1)
    }
]

### Relaciones

# realaciones de todas de las sentencias con los articulos
lista_relaciones = []

def relacion_articulos_sentencia(articulos, num):
    articulos_dict = {}
    for a,n in articulos:
        articulos_dict = {
            "sentencia_id": num,
            "articulo_id": n,
            "demanda":"inconstitucional"
        }
        lista_relaciones.append(articulos_dict)
    
    return lista_relaciones    

relacion_articulos_sentencia(articulos_318, 318)
relacion_articulos_sentencia(articulos_738, 738)
relacion_articulos_sentencia(articulos_782, 782)

## Save to csv

def save_to_csv(lista_diccionarios, nombre_archivo):
    # Obtener las claves del primer diccionario como encabezados
    keys = lista_diccionarios[0].keys()
    
    # Abrir el archivo en modo de escritura
    with open(nombre_archivo, 'w', newline='', encoding='utf-8') as output_file:
        dict_writer = csv.DictWriter(output_file, fieldnames=keys)
        dict_writer.writeheader()  # Escribir la cabecera del CSV
        dict_writer.writerows(lista_diccionarios)  # Escribir los datos del diccionario

    print(f"Archivo guardado como {nombre_archivo}")

save_to_csv(lista_sentencias, 'sentencias.csv')

save_to_csv(tutelas_list, 'tutelas.csv')

save_to_csv(lista_relaciones, 'relaciones_sentencias_articulos.csv')
