==> notebooks/03_app_chatbot.ipynb <==
import streamlit as st

from langchain.llms import HuggingFacePipeline
from langchain.memory import ConversationBufferMemory
from langchain.chains import ConversationalRetrievalChain
from langchain.prompts.prompt import PromptTemplate
from langchain.embeddings import HuggingFaceEmbeddings
from langchain.vectorstores import Neo4jVector
from transformers import AutoTokenizer, AutoModelForCausalLM, pipeline

from neo4j import GraphDatabase

# Configuración del modelo
MODEL_NAME = "google/gemma-7b"

# Test conection

class Neo4jConnection:

    def __init__(self, uri, user, password):
        self._driver = GraphDatabase.driver(uri, auth=(user, password))

    def close(self):
        self._driver.close()

    def query(self, query, parameters=None):
        with self._driver.session() as session:
            result = session.run(query, parameters)
            return result.data()

conn = Neo4jConnection(uri=NEO4J_URI, user=NEO4J_USER, password=NEO4J_PASSWORD)

query = "MATCH (n) RETURN n LIMIT 5"
results = conn.query(query)

for record in results:
    print(record)
