==> notebooks/04_langchain_gemma.ipynb <==
from dotenv import load_dotenv
import os
import re

from neo4j import GraphDatabase
from sentence_transformers import SentenceTransformer
from langchain.vectorstores import Neo4jVector

#Model import
import google.generativeai as genai


load_dotenv()

## Embeddings setup

URI = 'neo4j+s://295c5d0b.databases.neo4j.io'
NEO4J_USER = os.getenv('N_USER')
NEO4J_PASSWORD = os.getenv('N_PASSWORD')
G_KEY = os.getenv('GEMINI_API')

# Modelo optimizado para trabajar en espanol y para hacer preguntas.
model = SentenceTransformer('paraphrase-xlm-r-multilingual-v1')

# Se nesecita para crear los embedings en el nodo . Neo4j.Vector nesecita este el model_embedding y en este caso estoy utilizando el sentence_transformer
class SentenceEmbedding:
    def __init__(self, model):
        self.model = model

    def embed_query(self, text):
        return self.model.encode(text).tolist()  # Convierte a lista

    def embed_documents(self, texts):
        return self.model.encode(texts).tolist()  # Convierte a lista

# Crea una instancia de SentenceEmbedding
embedding = SentenceEmbedding(model)

# Esto solo se ejecuta una vez para crear los embedings en el nodo utilizando el texto de la propiedad "pregunta"
Neo4j_Vector_Store = Neo4jVector.from_existing_graph(
    embedding=embedding,
    url=URI,
    username=NEO4J_USER,
    password=NEO4J_PASSWORD,
    index_name="p_vector_index",
    node_label="preguntas",
    text_node_properties=["pregunta"],
    embedding_node_property="pregunta_embedding",
)

# Conectarnos a la based de datos para verificar las propiedades de los nodos y confirmar de la cracion de los embeddings.
# Conexión a la base de datos Neo4j
uri = "neo4j+s://295c5d0b.databases.neo4j.io"
driver = GraphDatabase.driver(uri, auth=(NEO4J_USER, NEO4J_PASSWORD))

# Hacer la consulta
def get_nodes(tx):
    query = """
    MATCH (n:preguntas) 
    RETURN n.pregunta_id AS id, n.pregunta_embedding AS embedding
    LIMIT 5    
    """
    result = tx.run(query)
    return [(record['id'], record['embedding']) for record in result]


with driver.session() as session:
    nodes = session.read_transaction(get_nodes)

print(nodes[:5])

## Load the model

genai.configure(api_key = G_KEY)

# New Gemini model Configurations
generation_config = {
  "temperature": 0.3,
  "top_p": 0.95,
  "top_k": 5,
  "max_output_tokens": 8192,
  "response_mime_type": "text/plain",
}

llm = genai.GenerativeModel(model_name='gemini-1.5-flash-latest', generation_config=generation_config)

## Functions

# Función para obtener nodos relacionados pregunta-[responde]->articulo
def get_related_nodes(session, pregunta_id):
    if not pregunta_id:
        raise ValueError("Pregunta_id no puede ser None o vacío.")    
    
    # Buscar en todos los nodos preguntas
    query = """
    MATCH (start:preguntas)   
    WHERE start.pregunta_id = $pregunta_id
    CALL apoc.path.expand(
        start,              // nodo inicial
        'responde>',        // tipo de relación
        '',                 // filtro de etiquetas (vacío para aceptar todas)
        1,                  // profundidad mínima
        1                   // profundidad máxima (antes era maxLevel)
    )
    YIELD path
    RETURN last(nodes(path)).articulo_name AS articulo_name,
           last(nodes(path)).contenido AS contenido_final
    """
    
    try:
        result = session.run(query, pregunta_id=pregunta_id)
        record = next(result) if result else None
        if record:
            text = f'articulo_numero: {record["articulo_name"]}\n{record["contenido_final"]}'
            return text  # o json.dumps(diccionario)
        return ""
    except Exception as e:
        print(f"Error ejecutando la consulta: {e}")
        return ""

## Test


query= 'Quienes pueden votar en una asamblea general'

vector_store = Neo4j_Vector_Store

results = vector_store.similarity_search(query, k=2)

results

print(results[0].metadata.get('pregunta_id'),
      results[1].metadata.get('pregunta_id'))

response_1 = get_related_nodes(session=driver.session(),pregunta_id=2504)

print(response_1)

response_2 = get_related_nodes(session=driver.session(),pregunta_id=2504)

print(response_2)

response = response_1+response_2

response

## Generar respuestas

my_prompt = f"Sos un asistente legal especializado en la ley 675 de Colombia y usando el siguiente contexto: {response_1} responde la siguiente pregunta:{query} de la manera mas detallada posible y si aplica utiliza algun ejemplo y al finalizar invita al usuario a volver a preguntar"

ans = llm.generate_content(my_prompt, generation_config=generation_config)

print(ans.text)
