==> notebooks/06_questions_generator.ipynb <==
# Load Data

import pandas as pd

df_1 = pd.read_csv(r'D:\0_Respaldo\000_KaggleX\Projects\Rag-Ley675\675_LAW\data\raw\ok_q\qq-0-132 .csv', encoding='latin-1')

df_1.info()

df_2 = pd.read_csv(r'D:\0_Respaldo\000_KaggleX\Projects\Rag-Ley675\675_LAW\data\raw\ok_q\qq-133-216 .csv', encoding='latin-1')

df_2.info()

df_3 = pd.read_csv(r'D:\0_Respaldo\000_KaggleX\Projects\Rag-Ley675\675_LAW\data\raw\ok_q\qq-217-910 (1).csv', encoding='latin-1')

df_3.info()

df_2.head(3)
