==> notebooks/07_Articles_Selection.ipynb <==
import pandas as pd

data = pd.read_csv(r'D:\0_Respaldo\000_KaggleX\Projects\Rag-Ley675\675_LAW\data\clean\articulos.csv')

data.head(1)

df = data.copy()

df.tail(5)

df['contenido'] = df['contenido'].astype(str)

df['conteo_palabras'] = df['contenido'].str.split().apply(len)

df['conteo_palabras'].describe()

# Definir los rangos para agrupar el conteo de palabras
bins = [0, 132, 216, 910]  # Puedes ajustar estos valores
labels = ['0-132', '131-216', '217-910']

# Crear la columna de rangos etáreos
df['rango_palabras'] = pd.cut(df['conteo_palabras'], bins=bins, labels=labels, right=True)


df.sample(5)

# Filtrar las filas que se encuentran en el rango '0-50'
df_rango_0_132 = df[df['rango_palabras'] == '0-132']

df_rango_131_216 = df[df['rango_palabras'] == '131-216']

df_rango_0_132

df_rango_217_910 = df[df['rango_palabras'] == '217-910']

df_rango_0_132.to_csv('articulos_0-132.csv', index=False)

df_rango_131_216

df_rango_131_216.to_csv('articulos_133-216.csv', index=False)

df_rango_131_216

df_rango_217_910

df_rango_217_910.to_csv('articulos_217-910.csv', index=False, encoding='Lating-1')

df_rango_217_910
