==> notebooks/08_question_integration.ipynb <==
import pandas as pd

data = pd.read_csv(r'D:\0_Respaldo\000_KaggleX\Projects\Rag-Ley675\675_LAW\data\clean\q-0-132.csv', encoding='latin-1')
data2 = pd.read_csv(r'D:\0_Respaldo\000_KaggleX\Projects\Rag-Ley675\675_LAW\data\clean\q-133-216.csv', encoding='latin-1')
data3 = pd.read_csv(r'D:\0_Respaldo\000_KaggleX\Projects\Rag-Ley675\675_LAW\data\clean\q-217-910.csv', encoding='latin-1')

data.head(13)

data2.tail(4)

data3.head(4)

df = pd.concat([data,data2], ignore_index=True)

df

df = pd.concat([df,data3], ignore_index=True) 

df_o = df.sort_values(by='articulo_id', ascending=True)

df_o

df1 = df_o[['pregunta_id', 'articulo_id']]

df1['responde'] = 'Hacia articulo'

df1

df1.to_csv('preguntas_articulos_enlace.csv', index=False)

### Crear nodo

df.to_csv(r'D:\0_Respaldo\000_KaggleX\Projects\Rag-Ley675\675_LAW\notebooks\preguntas_articulos_nodos.csv', index=False)
