==> notebooks/09_answer_generation.ipynb <==
import pandas as pd

data = pd.read_csv(r'D:\0_Respaldo\000_KaggleX\Projects\Rag-Ley675\675_LAW\notebooks\preguntas_articulos_enlace.csv')

df_data = pd.read_csv(r'D:\0_Respaldo\000_KaggleX\Projects\Rag-Ley675\675_LAW\notebooks\articulos_0-132.csv')

df_data.head(2)

data.head(3)

df = data.copy()

df_1 = df[df['articulo_id'] == 1]

df_1.head(3)
